requirements.txt
openai
beautifulsoup4
requests
newspaper3k
python-dotenv
streamlit
```